Add definition of AUTOMAKE in configure.amp
This is needed because it is used to rebuild the Makefile, and if
config.mk needs to be rebuilt as well, the definition of AUTOMAKE in
there will not be read.

For this to work, am__aux_dir is being set in dir.mk, a new include file
which is guaranteed to exist.