Remove config.status and configure as Makefile dependencies.
We had a problem that the Makefile symlink would be viewed as having the
modification date of the srcdir Makefile and would always be rebuilt as
being older than config.status, which didn't work when srcdir is
read-only. (So we had to do "make -L distcheck".)

We don't need them as dependencies anyway as all they do is
create a symbolic link and we know this has been done by the fact we're
running the file they created.