Further compression of template
Parent commit was buggy - there's no good way to disable a compound shell
command (e.g. with ; or &&) at configure time. Putting # inside a make
variable is impossible. The other way would be to have TWO make variables,
one containing "@: ", the other ": ".

Here we calculate depbase twice for the sake of less repetition in the
template.